requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic>=2.0
pydantic-settings>=2.0
pydantic[email]
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
celery==5.3.6
redis==5.0.1
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
jinja2>=3.1
websockets>=12.0
alembic>=1.13
aiosmtplib>=3.0
pytest>=8.0
httpx>=0.26
pytest-asyncio>=0.23
